ref: move navigation widget into own modules
